Makefile: remove static linking in Linux build
The removed static linking was for a desire to make rmc run
on systems where c libraries are not provided.

But it is nothing wrong to dynamically link rmc with libraries.
We also don't have to hardcode the way of linking in Makefile
since people can specify it in RMC_CFLAGS passed to make command.

Signed-off-by: Jianxun Zhang <jianxun.zhang@linux.intel.com>