refactoring work (APIs, build dependency, ... etc)
This is a major re-work in rmc with several key changes:

To query board-specific data in a rmc database, developers
now can take advantage of new defined APIs, either using
single-action APIs for a better performance of multiple
queries or simply calling a double-action API, which is
sufficient to perform a query in one step, to reduce the
footprint in code at client side. (Refer to inc/rmc_api.h.)

rmc tool itself is modified to use new APIs.

Memory map and allocated data can be released after use in user
space. In EFI context, rmc doesn't allocate memory for anything
returned.

In EFI build, necessary data types are defined in rmc in order
to get rid of dependency on any EFI implementation.

Miscellaneous fixes are also included when we move the pieces
around.

Signed-off-by: Jianxun Zhang <jianxun.zhang@linux.intel.com>